# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FyndMate API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SearchRequest {
  query: string;
  budget?: number;
  use_case?: string;
  platform?: string;
}

export interface Alternative {
  name: string;
  price: number;
  reason: string;
}

export interface PlatformPrice {
  price: number;
  url: string;
  affiliate_url?: string;
  delivery?: string;
}

export interface Prices {
  amazon: PlatformPrice;
  flipkart: PlatformPrice;
  best_deal: string;
  savings: number;
}

export type MainProductAffiliateLinks = { [key: string]: string };

export interface MainProduct {
  product_name: string;
  brand: string;
  estimated_price: number;
  match_score: number;
  pros: string[];
  cons: string[];
  explanation: string;
  amazon_search?: string;
  flipkart_search?: string;
  return_risk: string;
  return_risk_reason?: string;
  fake_review_warning: boolean;
  fake_review_reason?: string;
  alternatives?: Alternative[];
  prices?: Prices;
  affiliate_links?: MainProductAffiliateLinks;
}

export interface SearchResponse {
  success: boolean;
  main_product: MainProduct;
  prices?: Prices;
}

export interface GiftRequest {
  recipient: string;
  interests: string;
  budget?: number;
}

export interface Gift {
  name: string;
  price: number;
  reason: string;
}

export interface GiftResponse {
  success: boolean;
  gifts: Gift[];
}

export interface OutfitRequest {
  body_type?: string;
  occasion?: string;
  budget?: number;
  skin_tone?: string;
  style_preference?: string;
}

export interface OutfitItem {
  name: string;
  price: number;
}

export interface Outfit {
  name: string;
  occasion: string;
  total_price: number;
  items: OutfitItem[];
  color_palette?: string[];
  styling_tip: string;
  image_url?: string;
}

export interface OutfitResponse {
  success: boolean;
  outfits: Outfit[];
}

export interface PriceRequest {
  product_name: string;
  budget?: number;
}

export interface PriceResponse {
  success: boolean;
  amazon: PlatformPrice;
  flipkart: PlatformPrice;
  best_deal: string;
  savings: number;
}