# lib/api-zod/src/generated/types/gift.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FyndMate API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Gift {
  name: string;
  price: number;
  reason: string;
}