# lib/api-zod/src/generated/types/giftResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FyndMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Gift } from "./gift";

export interface GiftResponse {
  success: boolean;
  gifts: Gift[];
}