# lib/api-zod/src/generated/types/mainProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FyndMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Alternative } from "./alternative";
import type { MainProductAffiliateLinks } from "./mainProductAffiliateLinks";
import type { Prices } from "./prices";

export interface MainProduct {
  product_name: string;
  brand: string;
  estimated_price: number;
  match_score: number;
  pros: string[];
  cons: string[];
  explanation: string;
  amazon_search?: string;
  flipkart_search?: string;
  return_risk: string;
  return_risk_reason?: string;
  fake_review_warning: boolean;
  fake_review_reason?: string;
  alternatives?: Alternative[];
  prices?: Prices;
  affiliate_links?: MainProductAffiliateLinks;
}