# lib/api-zod/src/generated/types/outfit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FyndMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OutfitItem } from "./outfitItem";

export interface Outfit {
  name: string;
  occasion: string;
  total_price: number;
  items: OutfitItem[];
  color_palette?: string[];
  styling_tip: string;
  image_url?: string;
}