# lib/api-zod/src/generated/types/outfitRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FyndMate API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OutfitRequest {
  body_type?: string;
  occasion?: string;
  budget?: number;
  skin_tone?: string;
  style_preference?: string;
}