# lib/api-zod/src/generated/types/outfitResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FyndMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Outfit } from "./outfit";

export interface OutfitResponse {
  success: boolean;
  outfits: Outfit[];
}