# lib/api-zod/src/generated/types/priceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FyndMate API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PriceRequest {
  product_name: string;
  budget?: number;
}