# lib/api-zod/src/generated/types/prices.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FyndMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformPrice } from "./platformPrice";

export interface Prices {
  amazon: PlatformPrice;
  flipkart: PlatformPrice;
  best_deal: string;
  savings: number;
}