# lib/api-zod/src/generated/types/searchResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FyndMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MainProduct } from "./mainProduct";
import type { Prices } from "./prices";

export interface SearchResponse {
  success: boolean;
  main_product: MainProduct;
  prices?: Prices;
}